third_party/blink/web_tests/http/tests/security/isolatedWorld/isolated-world-eval-csp-expected.txt

CONSOLE MESSAGE: Testing main world. Eval should be blocked by main world CSP.
CONSOLE MESSAGE: EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'self'".

CONSOLE MESSAGE: PASS: eval blocked as expected.
CONSOLE MESSAGE: Testing isolated world with no csp. Eval should be allowed.
CONSOLE MESSAGE: PASS: eval allowed as expected.
CONSOLE MESSAGE: Testing isolated world with strict csp.
CONSOLE MESSAGE: EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'none'".

CONSOLE MESSAGE: PASS: eval blocked as expected.
CONSOLE MESSAGE: Testing isolated world with permissive csp.
CONSOLE MESSAGE: PASS: eval allowed as expected.
This tests the handling of unsafe-eval CSP checks and its interaction with the isolated world CSP.
